Improved networking and physics computations
To minimize the bandwidth consumption, only the acceleration of the ships
is sent to the clients. The clients compute the velocity and position
itself. IMPORTANT: acceleration/velocity/positions setters now have a
second argument to force synchronization with the clients. If it is not
set to true, the value will only be changed serverside.